refactor: removido os testes de dentro de minimal-API-generator e criado a pasta tests
